fix(formatter): avoid temporary files when formatting with RuboCop
By supplying the file content to RubyCop as a in-memory string the same
way as RuboCop's own CLI tool handles STDIN, we completely avoid having
to write any temporary files to disk, and formatting can occur purely
in-memory.

This also has the benefit of Solargraph no longer needing to be
responsible for finding and passing along any config file paths. RuboCop
will correctly find any relevant .rubocop.yml files for the file being
formatted, as RuboCop behaves exactly as if it was running against the
real file in it's original location on disk, except with the content it
is formatting coming from the in-memory string rather than reading it
from the file on disk.

A new blank formatter is also passed to RuboCop, just to make it print
as little as possible to STDOUT, since we don't need and ignore the
STDOUT produced.