feat(formatter): add "except" and "only" rubocop formatter options
They take a list or comma separated string of cops, which will be passed
to rubocop as --except and/or --only.

The "except" option is useful to disable cops which can cause annoyance
when using solargraph's formatter as a before save hook in text
editors.

I'm less aware of what the "only" option might be useful for, but might
as well add it so both are covered.