Fix query_symbols changed to exact match
I could not find anything about the lsp spec suggesting to use fuzzy matching for this function. Currently it breaks jumping to function in vim with `:tag` which is mapped to `workspace/symbol` in solargraph-lsp since all substring matches are also considered. This is how it works in `gopls` as well.

Issues that could be related #292