Remove accidental inclusion of Module's methods in objects (#884)
The instance methods on 'Module' aren't available on arbitrary
objects - don't include them in the api_map as such.

The inaccurate method list was causing problems at type-checking time
- when an external gem (RBS, in this case) had a method ("name") that
we didn't have a type for, it could look for the types on a method in
Method--in this case Method#name, which returns String.  Since RBS has
its own class ("TypeName"), this would fail to typecheck at :strict.

Also update a clip_spec test, which was expecting incorrect behavior
in completions.