[regression] Avoid non-deterministic behavior with better cache key (#917)
* Avoid non-deterministic behavior with better cache key

I've seen some dodgy non-repeatable behavior from Solargraph in the
last few days.  I was able to pin it down (so to speak) to a
particular use case I could repeat, and it seemed to be related to
links being cached by their words alone in Chain.

Unfortunately, I did not record enough about the scenario to reproduce
it :(

However, this change ought to strengthen the coherency of the links
part of our cache key without affecting performance and hopefully
squash this.

* Avoid non-deterministic behavior with better cache key

* Add Equality module to simplify use of equality_fields

* Drop source from Pin::Base#equality_fields

* Remove accidental addition

* Add additional fields to equality_fields

* Drop reference to future field