Fix pin combination consistency issue
I was hitting a strange 'works on my machine' issue and figured it out
that I wasn't seeing it because of a difference in where files lived
in CI vs my machine.  This resulted in different pin info being
selected during combination, as we were using the file location when
we had no better way to prefer one over the other.

This should make the result more consistent, making user and CI
pin-combination-triggered issues easier to reproduce.