[regression] Fix resolution of ambiguous argument types (#1098)
* [regression] Fix resolution of ambiguous argument types

This is a case where incorrect gates result in the wrong type being
used - found by strong typechecking on a branch ::Solargraph::Pin::Symbol
was resolved as ::Symbol in a generics scenario.

* Fix gates default value to match existing behavior

* Linting

* Linting

* Fix re-used variable name issue

* Fix pin combination consistency issue

I was hitting a strange 'works on my machine' issue and figured it out
that I wasn't seeing it because of a difference in where files lived
in CI vs my machine.  This resulted in different pin info being
selected during combination, as we were using the file location when
we had no better way to prefer one over the other.

This should make the result more consistent, making user and CI
pin-combination-triggered issues easier to reproduce.

* Bump for call.rb complexity